test: deployBountyContract means now deploy without BountyFactory
