+++ packages/network-contracts/test/hardhat/BrokerEconomics/Bounty.test.ts
@@ -14,7 +14,7 @@ import {
     TestContracts,
 } from "./deployTestContracts"
 
-import { deployBountyContract } from "./deployBountyContract"
+import { deployBountyContract } from "./deployBounty"
 
 describe("Bounty", (): void => {
     let admin: Wallet
@@ -46,17 +46,17 @@ describe("Bounty", (): void => {
         await (await token.transfer(broker.address, parseEther("100000"))).wait()
         await (await token.transfer(broker2.address, parseEther("100000"))).wait()
 
-        defaultBounty = await deployBountyContract(contracts, {skipBountyFactory: true})
+        defaultBounty = await deployBountyContract(contracts)
     })
 
     it("accepts 32 byte long address in transferAndCall data", async function(): Promise<void> {
-        const bounty = await deployBountyContract(contracts, {skipBountyFactory: true})
+        const bounty = await deployBountyContract(contracts)
         await (await token.transferAndCall(bounty.address, parseEther("1"), defaultAbiCoder.encode(["address"], [broker.address]))).wait()
         expect(await bounty.connect(broker).getMyStake()).to.be.equal(parseEther("1"))
     })
 
     it("accepts 2-step staking: approve + stake", async function(): Promise<void> {
-        const bounty = await deployBountyContract(contracts, {skipBountyFactory: true})
+        const bounty = await deployBountyContract(contracts)
         await (await token.approve(bounty.address, parseEther("1"))).wait()
         await (await bounty.stake(broker.address, parseEther("1"))).wait()
         expect(await bounty.connect(broker).getMyStake()).to.be.equal(parseEther("1"))
@@ -92,7 +92,7 @@ describe("Bounty", (): void => {
     })
 
     it("shows zero allocation after a withdraw", async function(): Promise<void> {
-        const bounty = await deployBountyContract(contracts, {skipBountyFactory: true})
+        const bounty = await deployBountyContract(contracts)
         await (await bounty.sponsor(parseEther("10000"))).wait()
         const start = await getBlockTimestamp()
 
@@ -110,7 +110,7 @@ describe("Bounty", (): void => {
     })
 
     it("shows zero allocation and zero stake after unstaking (no committed stake)", async function(): Promise<void> {
-        const bounty = await deployBountyContract(contracts, {skipBountyFactory: true})
+        const bounty = await deployBountyContract(contracts)
         await (await bounty.sponsor(parseEther("10000"))).wait()
         const start = await getBlockTimestamp()
 
@@ -154,27 +154,23 @@ describe("Bounty", (): void => {
     describe("IJoinPolicy negative tests", (): void => {
 
         it("error setting param on joinpolicy", async function(): Promise<void> {
-            await expect(deployBountyContract(contracts, {skipBountyFactory: true},
-                [testJoinPolicy], ["1"])) // 1 => TestJoinPolicy:setParam will revert
+            await expect(deployBountyContract(contracts, {}, [testJoinPolicy], ["1"])) // 1 => TestJoinPolicy:setParam will revert
                 .to.be.revertedWith("test-error: setting param join policy")
         })
 
         it("error setting param on joinpolicy no revert reason", async function(): Promise<void> {
-            await expect(deployBountyContract(contracts, {skipBountyFactory: true},
-                [testJoinPolicy], ["2"])) // 2 => TestJoinPolicy:setParam will revert without reason
+            await expect(deployBountyContract(contracts, {}, [testJoinPolicy], ["2"])) // 2 => TestJoinPolicy:setParam will revert without reason
                 .to.be.revertedWith("error_addJoinPolicyFailed")
         })
 
         it("error joining on joinpolicy", async function(): Promise<void> {
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true},
-                [testJoinPolicy], ["0"])
+            const bounty = await deployBountyContract(contracts, {}, [testJoinPolicy], ["0"])
             await expect(token.transferAndCall(bounty.address, 1, admin.address))
                 .to.be.revertedWith("test-error: onJoin join policy")
         })
 
         it("error joining on joinpolicy, empty error", async function(): Promise<void> {
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true},
-                [testJoinPolicy], ["0"])
+            const bounty = await deployBountyContract(contracts, {}, [testJoinPolicy], ["0"])
             await expect(token.transferAndCall(bounty.address, 2, admin.address))
                 .to.be.revertedWith("error_joinPolicyOnJoin")
         })
@@ -189,28 +185,28 @@ describe("Bounty", (): void => {
         })
 
         it("error onJoin on allocationPolicy", async function(): Promise<void> {
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true},
+            const bounty = await deployBountyContract(contracts, {},
                 [], [], testAllocationPolicy, "3") // 3 => onJoin will revert
             await expect(token.transferAndCall(bounty.address, parseEther("1"), admin.address))
                 .to.be.revertedWith("test_onJoin")
         })
 
         it("error onJoin on allocationPolicy, empty error", async function(): Promise<void> {
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true},
+            const bounty = await deployBountyContract(contracts, {},
                 [], [], testAllocationPolicy, "4") // 4 => onJoin will revert without reason
             await expect(token.transferAndCall(bounty.address, parseEther("1"), admin.address))
                 .to.be.revertedWith("error_allocationPolicyOnJoin")
         })
 
         it("error onleave on allocationPolicy", async function(): Promise<void> {
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true},
+            const bounty = await deployBountyContract(contracts, {},
                 [], [], testAllocationPolicy, "5") // 5 => onLeave will revert
             await (await token.transferAndCall(bounty.address, parseEther("1"), broker.address)).wait()
             await expect(bounty.connect(broker).unstake()).to.be.revertedWith("test_onLeave")
         })
 
         it("error onleave on allocationPolicy, empty error", async function(): Promise<void> {
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true},
+            const bounty = await deployBountyContract(contracts, {},
                 [], [], testAllocationPolicy, "6") // 6 => onLeave will revert without reason
             await (await token.transferAndCall(bounty.address, parseEther("1"), broker.address)).wait()
             await expect(bounty.connect(broker).unstake()).to.be.revertedWith("error_brokerLeaveFailed")
@@ -218,7 +214,7 @@ describe("Bounty", (): void => {
 
         it("error onStakeChange", async function(): Promise<void> {
             // 7 => onStakeChange will revert
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true}, [], [], testAllocationPolicy, "7")
+            const bounty = await deployBountyContract(contracts, {}, [], [], testAllocationPolicy, "7")
             await (await token.transferAndCall(bounty.address, parseEther("1"), admin.address)).wait()
             await expect(token.transferAndCall(bounty.address, parseEther("1"), admin.address))
                 .to.be.revertedWith("test_onStakeChange")
@@ -226,7 +222,7 @@ describe("Bounty", (): void => {
 
         it("error onStakeChange, empty error", async function(): Promise<void> {
             // 8 => onStakeChange revert without reason
-            const bounty = await deployBountyContract(contracts, {skipBountyFactory: true}, [], [], testAllocationPolicy, "8")
+            const bounty = await deployBountyContract(contracts, {}, [], [], testAllocationPolicy, "8")
             await (await token.transferAndCall(bounty.address, parseEther("1"), admin.address)).wait()
             await expect(token.transferAndCall(bounty.address, parseEther("1"), admin.address))
                 .to.be.revertedWith("error_stakeIncreaseFailed")
