+++ packages/network-contracts/test/hardhat/BrokerEconomics/BountyPolicies/DefaultLeavePolicy.test.ts
@@ -4,7 +4,7 @@ import { utils, Wallet } from "ethers"
 
 import { deployTestContracts, TestContracts } from "../deployTestContracts"
 import { advanceToTimestamp, getBlockTimestamp } from "../utils"
-import { deployBountyContract } from "../deployBountyContract"
+import { deployBountyContract } from "../deployBounty"
 
 const { parseEther, formatEther } = utils
 
@@ -41,8 +41,7 @@ describe("DefaultLeavePolicy", (): void => {
         const { token } = contracts
         const bounty = await deployBountyContract(contracts, {
             minBrokerCount: 2,
-            penaltyPeriodSeconds: 1000,
-            skipBountyFactory: true
+            penaltyPeriodSeconds: 1000
         })
         expect(!await bounty.isRunning())
         expect(!await bounty.isFunded())
@@ -88,7 +87,7 @@ describe("DefaultLeavePolicy", (): void => {
         // broker1:       400  +  300               =  700
         // broker2:               300  +  700       = 1000
         const { token } = contracts
-        const bounty = await deployBountyContract(contracts, { penaltyPeriodSeconds: 1000, skipBountyFactory: true })
+        const bounty = await deployBountyContract(contracts, { penaltyPeriodSeconds: 1000 })
         expect(!await bounty.isRunning())
         expect(!await bounty.isFunded())
